requirements.txt
picamera
websocket-client
